+++ src/syncData.ts
@@ -85,10 +85,10 @@ export const initSync = (
   const processChangeStreamRecords = async (docs: ChangeStreamDocument[]) => {
     try {
       const operations = []
-      const counts = getInitialCounts()
+      const operationCounts = getInitialCounts()
       for (const doc of docs) {
-        counts[doc.operationType]++
         if (doc.operationType === 'insert') {
+          operationCounts[doc.operationType]++
           operations.push([
             { create: { _index: index, _id: doc.fullDocument._id.toString() } },
             mapper(doc.fullDocument),
@@ -97,12 +97,14 @@ export const initSync = (
           doc.operationType === 'update' ||
           doc.operationType === 'replace'
         ) {
+          operationCounts[doc.operationType]++
           const document = doc.fullDocument ? mapper(doc.fullDocument) : {}
           operations.push([
             { index: { _index: index, _id: doc.documentKey._id.toString() } },
             document,
           ])
         } else if (doc.operationType === 'delete') {
+          operationCounts[doc.operationType]++
           operations.push([
             { delete: { _index: index, _id: doc.documentKey._id.toString() } },
           ])
@@ -120,10 +122,14 @@ export const initSync = (
           fail: numErrors,
           errors,
           changeStream: true,
-          counts,
+          operationCounts,
         })
       } else {
-        emit('process', { success: docs.length, changeStream: true, counts })
+        emit('process', {
+          success: docs.length,
+          changeStream: true,
+          operationCounts,
+        })
       }
     } catch (e) {
       emit('error', { error: e, changeStream: true })
@@ -133,6 +139,7 @@ export const initSync = (
    * Process initial scan documents.
    */
   const processRecords = async (docs: ChangeStreamInsertDocument[]) => {
+    const operationCounts = { insert: docs.length }
     try {
       const response = await elastic.bulk({
         operations: docs.flatMap((doc) => [
@@ -149,9 +156,14 @@ export const initSync = (
           fail: numErrors,
           errors,
           initialScan: true,
+          operationCounts,
         })
       } else {
-        emit('process', { success: docs.length, initialScan: true })
+        emit('process', {
+          success: docs.length,
+          initialScan: true,
+          operationCounts,
+        })
       }
     } catch (e) {
       emit('error', { error: e, initialScan: true })